user controller: migrate relevant routes to /signup
